docs&refactor(wzn): improve docstrings and move image&video handling methods in `MultimodalDataProcessor` to new folders
- update docstring to align with newest format
- add explanation added for label and removed unused variable and import in rft_dataset
- moved to_pil, normalize_images and get_images_num to lightrft/trainer/image_utils.py. And moved frames_to_tensor, single_video_to_tensor, to_video_tensor, normalize_videos and get_videos_num to lightrft/trainer/image_utils.py

+++ lightrft/trainer/ppo_trainer_vl.py
@@ -850,7 +850,7 @@ class PPOTrainerVL(ABC):
                 for sub_k, sub_v in v.items():
                     log_key = f"{k}/{sub_k}"
                     if isinstance(sub_v, torch.Tensor):
-                        status[log_key] = sub_v.float().mean().item()
+                        status[log_key] = sub_v.mean().item()
                     elif isinstance(sub_v, list) and sub_v and isinstance(sub_v[0], (int, float)):
                         status[log_key] = sum(sub_v) / len(sub_v)
                     elif isinstance(sub_v, (int, float)):
